addded competitive page

+++ src/main/java/com/harshit/AuraTracker/Service/GamificationService.java
@@ -3,5 +3,5 @@ package com.harshit.AuraTracker.Service;
 import com.harshit.AuraTracker.modal.Student;
 
 public interface GamificationService {
-    public Student addPoints(int studentId, int points) throws Exception;
+    public Student addPoints(Long studentId, int points) throws Exception;
 }
